[FEATURE][KED-2699] `AbstractConfigLoader` with delegated `_get_config_from_patterns` functionality (#1185)
